Removal of Pages
Little value, and added complexity and point of failures, seem the only purpose of the 'Pages' config.  Ditched this so we just have Header and Lines.  Also reduces the amount of config required.  We already tag page numbers on to transaction records, but we'll lose the continuous 'sheet' numbers that HSBC use.  Can always be re-introducred later if missed.